+++ Tests/ParseSwiftTests/IOS13Tests.swift
@@ -76,13 +76,11 @@ class IOS13Tests: XCTestCase { // swiftlint:disable:this type_body_length
         #endif
         try ParseStorage.shared.deleteAll()
 
-        guard let fileManager = ParseFileManager(),
-              let defaultDirectoryPath = fileManager.defaultDataDirectoryPath else {
+        guard let fileManager = ParseFileManager() else {
             throw ParseError(code: .unknownError, message: "Should have initialized file manage")
         }
 
-        let directory2 = defaultDirectoryPath
-            .appendingPathComponent(ParseConstants.fileDownloadsDirectory, isDirectory: true)
+        let directory2 = try ParseFileManager.downloadDirectory()
         let expectation2 = XCTestExpectation(description: "Delete files2")
         fileManager.removeDirectoryContents(directory2) { _ in
             expectation2.fulfill()